refactor(app): introduce FSD pages layer and thin route files
Move whole-screen composition into a dedicated pages layer and reduce
each route file to a thin adapter that bridges route params to its page.
Rename the stats route and screen to analytics.
