+++ src/pages/LoginPage.tsx
@@ -1,0 +1,3 @@
+import { LoginForm } from '@/features/auth';
+
+export const LoginPage = () => <LoginForm />;
