+++ src/pages/NewQuizPage.tsx
@@ -1,0 +1,3 @@
+import { CreateQuizForm } from '@/features/quiz';
+
+export const NewQuizPage = () => <CreateQuizForm />;
