+++ src/pages/RegisterPage.tsx
@@ -1,0 +1,3 @@
+import { RegisterForm } from '@/features/auth';
+
+export const RegisterPage = () => <RegisterForm />;
