+++ src/routes/(app)/_appLayout.history.tsx
@@ -1,18 +1,5 @@
 import { createFileRoute } from '@tanstack/react-router';
-import { GameHistoryList } from '@/features/user';
-
-const HistoryPage = () => {
-  return (
-    <div className='flex flex-col gap-6'>
-      <div>
-        <h1 className='text-2xl font-bold'>Game History</h1>
-        <p className='text-muted-foreground'>Your past games as host and player.</p>
-      </div>
-
-      <GameHistoryList />
-    </div>
-  );
-};
+import { HistoryPage } from '@/pages/HistoryPage';
 
 export const Route = createFileRoute('/(app)/_appLayout/history')({
   head: () => ({
