+++ src/routes/auth/_authLayout.login.tsx
@@ -1,9 +1,9 @@
 import { createFileRoute } from '@tanstack/react-router';
-import { LoginForm } from '@/features/auth';
+import { LoginPage } from '@/pages/LoginPage';
 
 export const Route = createFileRoute('/auth/_authLayout/login')({
   head: () => ({
     meta: [{ title: 'AskIt ⋅ Log in' }],
   }),
-  component: LoginForm,
+  component: LoginPage,
 });
